+++ frontend/src/components/atoms/LikeButton.tsx
@@ -1,62 +1,9 @@
-import { FC, useState, useEffect } from 'react';
+import { FC } from 'react';
 import { Button, Icon } from 'semantic-ui-react';
-import { FetchLikeCreate, FetchLikeDestroy, FetchLikeExists } from 'apis/Like';
-import { useParams, useHistory } from 'react-router-dom';
+import useLikeButton from 'hooks/LikeButton';
 
 const LikedButton: FC = () => {
-  const [state, setState] = useState({ liked: false, count: 0 });
-  const { id } = useParams<{ id: string }>();
-  const history = useHistory();
-
-  useEffect(() => {
-    FetchLikeExists(id)
-      .then((result) => {
-        if (result !== 0 && result.liked) {
-          setState((prev) => ({ ...prev, ...result }));
-        } else if (result !== 0 && !result.liked) {
-          setState((prev) => ({ ...prev, ...result }));
-        } else {
-          history.push(`/product/${id}`, {
-            message: 'エラーが発生しました。',
-            type: 'error',
-          });
-        }
-      })
-      .catch(() =>
-        history.push(`/`, {
-          message: 'エラーが発生しました。',
-          type: 'error',
-        }),
-      );
-  }, [id, history]);
-
-  const onCreate = () =>
-    FetchLikeCreate(id).then((result) =>
-      result === 201
-        ? setState((prevState) => ({
-            ...prevState,
-            liked: true,
-            count: prevState.count + 1,
-          }))
-        : history.push(`/product/${id}`, {
-            message: 'エラーが発生しました。',
-            type: 'error',
-          }),
-    );
-
-  const onDestroy = () =>
-    FetchLikeDestroy(id).then((result) =>
-      result === 201
-        ? setState((prevState) => ({
-            ...prevState,
-            liked: false,
-            count: prevState.count - 1,
-          }))
-        : history.push(`/product/${id}`, {
-            message: 'エラーが発生しました。',
-            type: 'error',
-          }),
-    );
+  const { state, onCreate, onDestroy } = useLikeButton();
 
   return state.liked ? (
     <Button
